### 2024-03-14 8d6e55c
Fix doi's for JOSS draft.
