+++ doc/JOSS/1.0/paper.bib
@@ -18,7 +18,7 @@ volume = {20},
 number = {3},
 pages = {1665-1675},
 keywords = {3-D modeling, numerical simulation, MATLAB toolbox, 3-D temperature distribution},
-doi = {https://doi.org/10.1029/2018GC008057},
+doi = {10.1029/2018GC008057},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2018GC008057},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2018GC008057},
 abstract = {Abstract Creating the initial geometry and temperature configuration of 3-D numerical simulations is a challenging task. Professional tools are expensive. They often have a steep learning curve and do mostly not interface with the numerical simulation software used by the geodynamics and tectonics academic community. There, we developed geometry Input/Output (geomIO), a MATLAB toolbox to create the initial configuration of geological models regarding model geometry and temperature structure. geomIO allows users to create a geo-referenced 3-D volume by drawing multiple 2-D cross sections in a standard vector graphics editor. The volume is then used to assign material properties and set up initial temperature distribution on a set of vertices. In 2-D mode, polygons can also be used to create a triangular mesh. In addition to the standard functionality, the gravity anomaly of any geometry created with geomIO can be calculated. In this paper, we give an overview of the basic functionalities of geomIO. We illustrate the strength of our software with advanced tectonic and geodynamic applications that could not have been performed with currently available free software. Applications include the Himalayan orogen, the Japanese subduction zones, present-day salt diapirs, and small-scale tectonic structures. We describe algorithms and file formats in the supporting information. The toolbox is user-friendly and flexible. Users can use custom pipelines or preset data processing pipelines, so most applications require only a few lines of code. geomIO is distributed under the GNU General Public License and includes an online wiki with tutorials and additional examples.},
@@ -33,7 +33,7 @@ volume = {23},
 number = {6},
 pages = {e2021GC010265},
 keywords = {geometry parameterization, variable geometry, salt tectonics, subduction velocity, inversion, subduction angle},
-doi = {https://doi.org/10.1029/2021GC010265},
+doi = {10.1029/2021GC010265},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2021GC010265},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2021GC010265},
 note = {e2021GC010265 2021GC010265},
@@ -61,6 +61,7 @@ journal = {Pure appl. geophys.},
 volume = {159},
 pages = {2335–2356},
 year = {2002},
+doi = {10.1007/s00024-002-8738-3},
 url = {https://doi.org/10.1007/s00024-002-8738-3}
 }
 
@@ -155,7 +156,7 @@ DOI = {10.5194/gmd-14-3899-2021}
   pages = {118471},
   year = {2024},
   issn = {0012-821X},
-  doi = {https://doi.org/10.1016/j.epsl.2023.118471},
+  doi = {10.1016/j.epsl.2023.118471},
   url = {https://www.sciencedirect.com/science/article/pii/S0012821X23004843},
   author = {Erik {van der Wiel} and Douwe J.J. {van Hinsbergen} and Cedric Thieulot and Wim Spakman},
   keywords = {Mantle convection, Slab sinking rates, Mantle mixing, Modelling constraints},
@@ -196,7 +197,7 @@ DOI = {10.5194/gmd-14-3899-2021}
   number = {8},
   pages = {e2022JB025877},
[…]
