+++ doc/JOSS/1.0/paper.bib
@@ -98,7 +98,7 @@ DOI = {10.5194/gmd-14-3899-2021}
   year = "1970",
   note = "Geothermal Problems",
   issn = "0040-1951",
-  doi = "https://doi.org/10.1016/0040-1951(70)90115-0",
+  doi = "10.1016/0040-1951(70)90115-0",
   url = "http://www.sciencedirect.com/science/article/pii/0040195170901150",
   author = "D.P. McKenzie"
 }
