### 2024-03-14 #6470 @MFraters
Dear @kthyng, thank you for all the information!

> Also, take a look at the DOI notes above.

I have fixed the doi's. There is still one doi missing, because the work it references is an AGU (https://www.agu.org/fall-meeting) abstract. When the paper is submitted I can change the reference.

> Also @MFraters in your previous publication was there a software review as part of the paper review?

There is a paper about version 0.2.0 of the code, which you can find here, including the reviews: https://se.copernicus.org/articles/10/1785/2019/. These where reviews on the paper, not on the code.

> In the meantime, please suggest 5 reviewers from the database listed above or your own (non-conflicted) extended network. Their github handles are most useful to receive but please don't use "@" to reference them since it will prematurely ping them.

Here are my suggestions for 5 reviewers from the database: busstoptaktik, cpgr, krischer, leouieda, SaumikDana. If you need more suggestions, please let me know.
