### 2024-04-23 #6671 @editorialbot
[REVIEW]: The Geodynamic World Builder: A planetary structure creator for the geosciences
**Submitting author:** @mfraters (<a href="http://orcid.org/0000-0003-0035-7723">Menno Fraters</a>)
**Repository:** https://github.com/GeodynamicWorldBuilder/WorldBuilder/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
