+++ .github/workflows/test.yml
@@ -65,7 +65,7 @@ jobs:
     strategy: 
       fail-fast: false
       matrix: 
-        os: ['ubuntu-latest','macos-latest','macos-11']
+        os: ['ubuntu-latest','macos-12','macos-11']
         build_type: ['Release', 'Debug']
       
     runs-on: ${{ matrix.os }}
@@ -92,7 +92,7 @@ jobs:
             echo "3"
             ls /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/libexec
             echo "4"
-            sudo xcode-select --switch /Applications/Xcode_15.0.1.app
+            sudo xcode-select --switch /Applications/Xcode_13.2.1.app
          fi
       shell: bash
       
@@ -214,7 +214,7 @@ jobs:
     strategy: 
       fail-fast: false
       matrix: 
-        os: ['ubuntu-latest','macos-latest']
+        os: ['ubuntu-latest','macos-12']
         build_type: ['Release']
       
     runs-on: ${{ matrix.os }}
@@ -230,7 +230,7 @@ jobs:
          elif [ "$RUNNER_OS" == "macOS" ]; then
             sudo brew install open-mpi || true    
             echo "4"
-            sudo xcode-select --switch /Applications/Xcode_15.0.1.app
+            sudo xcode-select --switch /Applications/Xcode_13.2.1.app
          else          
             echo "$RUNNER_OS not supported"
             exit 1
