### 2024-05-02 4d30f92
fix manual toctree entry.
