+++ cookbooks/index.md
@@ -12,5 +12,4 @@ This section contains self-contained cookbooks on how to design different geodyn
 3d_cartesian_transform_fault/doc/README
 simple_subduction_2d_cartesian/doc/README
 simple_subduction_2d_chunk/doc/README
-
 ```
