### 2024-05-08 #6671 @busstoptaktik
I have serious trouble compiling the software (using cmake, ninja, and a recent gcc), so comparing the paper and the actual software has proven tricky. Will submit an issue to the software repo, but busy the next 5 days, so that will be next week.
