+++ .github/workflows/test.yml
@@ -65,7 +65,7 @@ jobs:
     strategy: 
       fail-fast: false
       matrix: 
-        os: ['ubuntu-latest','macos-12','macos-11']
+        os: ['ubuntu-latest','macos-12','macos-11','macos-13']
         build_type: ['Release', 'Debug']
       
     runs-on: ${{ matrix.os }}
@@ -92,7 +92,11 @@ jobs:
             echo "3"
             ls /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/usr/libexec
             echo "4"
-            sudo xcode-select --switch /Applications/Xcode_13.2.1.app
+            if [ "${{ matrix.os }}" == "macos-13" ]; then
+                sudo xcode-select --switch /Applications/Xcode_15.0.1.app
+            else
+                sudo xcode-select --switch /Applications/Xcode_13.2.1.app
+            fi
          fi
       shell: bash
       
@@ -214,7 +218,7 @@ jobs:
     strategy: 
       fail-fast: false
       matrix: 
-        os: ['ubuntu-latest','macos-12']
+        os: ['ubuntu-latest','macos-12','macos-13']
         build_type: ['Release']
       
     runs-on: ${{ matrix.os }}
@@ -230,7 +234,11 @@ jobs:
          elif [ "$RUNNER_OS" == "macOS" ]; then
             sudo brew install open-mpi || true    
             echo "4"
-            sudo xcode-select --switch /Applications/Xcode_13.2.1.app
+            if [ "${{ matrix.os }}" == "macos-13" ]; then
+                sudo xcode-select --switch /Applications/Xcode_15.0.1.app
+            else
+                sudo xcode-select --switch /Applications/Xcode_13.2.1.app
+            fi
          else          
             echo "$RUNNER_OS not supported"
             exit 1
