### 2024-05-09 dc585f2
Update and fix most ubuntu testers to 24.04.
