### 2024-05-09 #6671 @MFraters
@busstoptaktik, I am sorry to hear that you have issues with compiling the code. If you would be able post your operating system and your gcc version (and cmake and ninja versions if not too much trouble) before that time I can try to reproduce the issue based on that in the meantime. If not, I of course completely understand and then I am looking forward to your issue in the software repository next week.
