+++ source/world_builder/features/continental_plate_models/grains/random_uniform_distribution_deflected.cc
@@ -29,6 +29,7 @@
 #include "world_builder/types/one_of.h"
 #include "world_builder/types/unsigned_int.h"
 #include "world_builder/types/value_at_points.h"
+#include "world_builder/utilities.h"
 #include "world_builder/world.h"
 
 namespace WorldBuilder
@@ -89,7 +90,11 @@ namespace WorldBuilder
                             Types::Array(Types::Double(1),0),
                             "A list of the deflections of all of the grains in each composition between 0 and 1.");
 
+          prm.declare_entry("basis rotation matrices", Types::Array(Types::Array(Types::Array(Types::Double(0),3,3),3,3),0),
+                            "A list with the rotation matrices of the grains which are present there for each compositions.");
 
+          prm.declare_entry("basis Euler angles z-x-z", Types::Array(Types::Array(Types::Double(0),3,3),0),
+                            "A list with the z-x-z Euler angles of the grains which are present there for each compositions.");
 
         }
 
@@ -102,11 +107,36 @@ namespace WorldBuilder
           max_depth = max_depth_surface.maximum;
           compositions = prm.get_vector<unsigned int>("compositions");
 
+          const bool set_euler_angles = prm.check_entry("basis Euler angles z-x-z");
+          const bool set_rotation_matrices = prm.check_entry("basis rotation matrices");
+
+          WBAssertThrow(!(set_euler_angles == true && set_rotation_matrices == true),
+                        "Only Euler angles or Rotation matrices may be set, but both are set for " << prm.get_full_json_path());
+
+
+          WBAssertThrow(!(set_euler_angles == false && set_rotation_matrices == false),
+                        "Euler angles or Rotation matrices have to be set, but neither are set for " << prm.get_full_json_path());
+
+          if (set_euler_angles)
+            {
+              std::vector<std::array<double,3> > basis_euler_angles_vector = prm.get_vector<std::array<double,3> >("basis Euler angles z-x-z");
+              basis_rotation_matrices.resize(basis_euler_angles_vector.size());
+              for (size_t i = 0; i<basis_euler_angles_vector.size(); ++i)
+                {
+                  basis_rotation_matrices[i] = Utilities::euler_angles_to_rotation_matrix(basis_euler_angles_vector[i][0],basis_euler_angles_vector[i][1],basis_euler_angles_vector[i][2]);
+                }
+
+            }
+          else
+            {
+              basis_rotation_matrices = prm.get_vector<std::array<std::array<double,3>,3> >("basis rotation matrices");
+            }
+
           operation = prm.get<std::string>("orientation operation");
           grain_sizes = prm.get_vector<double>("grain sizes");
           normalize_grain_sizes = prm.get_vector<bool>("normalize grain sizes");
           deflections = prm.get_vector<double>("deflections");
-
+          
           WBAssertThrow(compositions.size() == grain_sizes.size(),
                         "There are not the same amount of compositions (" << compositions.size()
                         << ") and grain_sizes (" << grain_sizes.size() << ").");
@@ -116,6 +146,28 @@ namespace WorldBuilder
           WBAssertThrow(compositions.size() == deflections.size(),
                         "There are not the same amount of compositions (" << compositions.size()
                         << ") and deflections (" << deflections.size() << ").");
+          WBAssertThrow(compositions.size() == basis_rotation_matrices.size(),
+                        "There are not the same amount of compositions (" << compositions.size()
+                        << ") and rotation_matrices (" << basis_rotation_matrices.size() << ").");
+        }
+
+
+        std::array<std::array<double,3>,3>
[…]
