### 2024-05-30 b10424a
Simplify cmake C++ version setting
