### 2024-05-30 fc4bd46
Update CMakeLists.txt
Co-authored-by: Timo Heister <timo.heister@gmail.com>
