+++ CMakeLists.txt
@@ -4,7 +4,7 @@ project(WorldBuilder C CXX)
 
 set(CMAKE_CXX_STANDARD 14)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-set (CMAKE_CXX_EXTENSIONS OFF)
+set(CMAKE_CXX_EXTENSIONS OFF)
 
 # SWIG: use standard target name.
 if(POLICY CMP0078)
