### 2024-05-30 c576837
Move matrix multiplication to utilities and add unit test
Fix indentation
