+++ source/world_builder/features/plume_models/grains/random_uniform_distribution_deflected.cc
@@ -152,136 +152,119 @@ namespace WorldBuilder
         }
 
 
-        std::array<std::array<double,3>,3>
-        RandomUniformDistributionDeflected::matrix_multiply(const std::array<std::array<double,3>,3> mat1, const std::array<std::array<double,3>,3> mat2) const
-        {
-          std::array<std::array<double,3>,3> result;
-          for (int i = 0; i < 3; i++)
-            {
-              for (int j = 0; j < 3; j++)
-                {
-                  result[i][j] = 0;
-                  for (int k = 0; k < 3; k++)
-                    {
-                      result[i][j] += mat1[i][k] * mat2[k][j];
-                    }
-                }
-            }
-          return result;
-        }
 
 
         WorldBuilder::grains
         RandomUniformDistributionDeflected::get_grains(const Point<3> & /*position_in_cartesian_coordinates*/,
-                            const Objects::NaturalCoordinate & /*position_in_natural_coordinates*/,
-                            const double depth,
-                            const unsigned int composition_number,
-                            WorldBuilder::grains grains_,
-                            const double  /*feature_min_depth*/,
-                            const double /*feature_max_depth*/) const
+                                                       const Objects::NaturalCoordinate & /*position_in_natural_coordinates*/,
+                                                       const double depth,
+                                                       const unsigned int composition_number,
+                                                       WorldBuilder::grains grains_,
+                                                       const double  /*feature_min_depth*/,
+                                                       const double /*feature_max_depth*/) const
         {
           WorldBuilder::grains  grains_local = grains_;
           if (depth <= max_depth && depth >= min_depth)
             {
               for (unsigned int i =0; i < compositions.size(); ++i)
                 {
-                    if (compositions[i] == composition_number)
+                  if (compositions[i] == composition_number)
                     {
-                        std::uniform_real_distribution<> dist(0.0,1.0);
-                        for (auto &&it_rotation_matrices : grains_local.rotation_matrices)
+                      std::uniform_real_distribution<> dist(0.0,1.0);
+                      for (auto &&it_rotation_matrices : grains_local.rotation_matrices)
                         {
-                            // set a uniform random a_cosine_matrix per grain
-                            // This function is based on an article in Graphic Gems III, written by James Arvo, Cornell University (p 116-120).
-                            // The original code can be found on  http://www.realtimerendering.com/resources/GraphicsGems/gemsiii/rand_rotation.c
-                            // and is licenend accourding to this website with the following licence:
-                            //
-                            // "The Graphics Gems code is copyright-protected. In other words, you cannot claim the text of the code as your own and
-                            // resell it. Using the code is permitted in any program, product, or library, non-commercial or commercial. Giving credit
-                            // is not required, though is a nice gesture. The code comes as-is, and if there are any flaws or problems with any Gems
-                            // code, nobody involved with Gems - authors, editors, publishers, or webmasters - are to be held responsible. Basically,
-                            // don't be a jerk, and remember that anything free comes with no guarantee.""
-                            //
-                            // The book saids in the preface the following: "As in the first two volumes, all of the C and C++ code in this book is in
-                            // the public domain, and is yours to study, modify, and use."
+                          // set a uniform random a_cosine_matrix per grain
+                          // This function is based on an article in Graphic Gems III, written by James Arvo, Cornell University (p 116-120).
+                          // The original code can be found on  http://www.realtimerendering.com/resources/GraphicsGems/gemsiii/rand_rotation.c
+                          // and is licenend accourding to this website with the following licence:
[…]
