+++ include/world_builder/features/fault_models/grains/random_uniform_distribution_deflected.h
@@ -80,15 +80,7 @@ namespace WorldBuilder
              * class
              */
             void parse_entries(Parameters &prm) override final;
-
-            /*
-             * Returns the result of the multiplication of two 3*3 matrix,
-             * used in applying the random uniform distribution rotation matrix
-             * to a given orientation (rotation matrix)
-             */
-            std::array<std::array<double,3>,3>
-            matrix_multiply(const std::array<std::array<double,3>,3> mat1, std::array<std::array<double,3>,3> const mat2) const;
-
+            
             /**
              * Returns a grains based on the given position, composition (e.g.
              * olivine and/or enstatite)depth in the model, gravity and current grains.
