+++ include/world_builder/features/oceanic_plate_models/grains/random_uniform_distribution_deflected.h
@@ -81,14 +81,6 @@ namespace WorldBuilder
              */
             void parse_entries(Parameters &prm, const std::vector<Point<2>> &coordinates) override final;
 
-            /*
-             * Returns the result of the multiplication of two 3*3 matrix,
-             * used in applying the random uniform distribution rotation matrix
-             * to a given orientation (rotation matrix)
-             */
-            std::array<std::array<double,3>,3>
-            matrix_multiply(const std::array<std::array<double,3>,3> mat1, std::array<std::array<double,3>,3> const mat2) const;
-
             /**
              * Returns a grains based on the given position, composition (e.g.
              * olivine and/or enstatite)depth in the model, gravity and current grains.

[…]
