### 2024-06-01 4ad7337
Remove std::transform as a c++17 feature
