+++ source/world_builder/utilities.cc
@@ -1521,7 +1521,7 @@ namespace WorldBuilder
                 }
             }
         }
- 
+
       return result;
     }
   } // namespace Utilities
