+++ source/world_builder/features/oceanic_plate_models/grains/random_uniform_distribution_deflected.cc
@@ -266,9 +266,10 @@ namespace WorldBuilder
                               // std::transform(grains_local.sizes.begin(), grains_local.sizes.end(), grains_local.sizes.begin(),
                               //                [one_over_total_size](double sizes) -> double { return sizes *one_over_total_size; });
                               // Apply the transformation using a loop
-                              for (auto &&size : grains_local.sizes) {
+                              for (auto &&size : grains_local.sizes)
+                                {
                                   size = size*one_over_total_size;
-                              }
+                                }
                             }
 
                           return grains_local;
