### 2024-06-01 b67e233
Update 08_passive_margin_variable_depth.md
This is the first commit to fix only typos and grammatical errors.

This will be followed by a second commit to modify the paragraph we discussed.
