+++ doc/sphinx/user_manual/basic_starter_tutorial/08_passive_margin_variable_depth.md
@@ -2,7 +2,7 @@
 Adding a passive margin with variable depth
 ============================
 
-Although we have hinted towards it before, there is also an [area feature](part:user_manual:chap:concepts:sec:area_features) called `continental plate`. It works as you would expect based on what you have seen with the oceanic plate, but we are going to use it to do something very cool in our model: have variable depths of a feature or even of layers within a given feature!
+Although we have hinted towards it before, there is also an [area feature](part:user_manual:chap:concepts:sec:area_features) called `continental plate`. It works as you would expect based on what you have seen with the oceanic plate. Here we are going to use it to do something very cool in our model - specify a feature with variable depths and layers within a given feature!
 
 To showcase this, we will be adding a passive margin to our model at the overriding plate side of our model. To start out, we need to know that both the `min depth` and `max depth` accept two types of values: a number and an array of values at points. 
 
@@ -10,7 +10,7 @@ To showcase this, we will be adding a passive margin to our model at the overrid
 For an example of what this looks like in technical terms, see {ref}`open_features_items_oneOf_1_max-depth`.
 ```
 
-The number input we have seen before and it sets the maximum depth of the feature to a single value. The value at points system works a bit differently and can be used to achieve the same result. To start, this is exactly what we are going to to for the `max depth` of the continental feature. 
+The number input we have seen before and it sets the maximum depth of the feature to a single value. The value at points system works a bit differently and can be used to achieve the same result. To start, this is exactly what we are going to do for the `max depth` of the continental feature.
 
 The value at points system creates a list of points that includes all the initial edge points of an object and assigns a value to them. We then interpolate between close points to obtain the actual values at the remaining points according to the model type, e.g., linearly to the points set in `max depth` below.
 
@@ -86,5 +86,5 @@ If you provide a point twice, explicitly, or implicitly through the use of the c
 :alt: Basic Starter Tutorial section 8. 
 :align: center
 
-Basic Starter Tutorial section 8. The top part of the figure shows where the composition as been assigned as an object. Currently it shows composition 0 as green, composition 1 as yellow and composition 3 as blue. The bottom part shows the temperature as seen slightly from below where only temperatures between 300K and 1600K are shown. The added continental plate with variable thickness of it's two layers is now visible on the left side of the image.
+Basic Starter Tutorial section 8. The top part of the figure shows where the composition has been assigned as an object. Currently it shows composition 0 as green, composition 1 as yellow and composition 3 as blue. The bottom part shows the temperature as seen slightly from below where only temperatures between 300K and 1600K are shown. The added continental plate with variable thickness of its two layers is now visible on the left side of the image.
 ```
