### 2024-06-01 11a69c4
Update 18_2D_models.md
Fixes typos. 

MAJOR edit to description of first figure. CHECK.

Also, in the first note, is the following correct:
 ... the origin of the  2D cross section is located at [0,200e3] ...

Should it be 450e3 ?
