### 2024-06-04 be0e099
Fix floating-point comparison with approx
