### 2024-06-05 4e5288b
Add linux clang tester.
