### 2024-06-05 b044567
Fix plume tests, address comments
Remove unnessary comments and lines
