+++ include/world_builder/features/continental_plate_models/grains/random_uniform_distribution_deflected.h
@@ -81,8 +81,6 @@ namespace WorldBuilder
              */
             void parse_entries(Parameters &prm, const std::vector<Point<2>> &coordinates) override final;
 
-
-
             /**
              * Returns a grains based on the given position, composition (e.g.
              * olivine and/or enstatite)depth in the model, gravity and current grains.
