### 2024-06-05 87022ec
Add more flags to cmakelists.
