### 2024-06-05 63fd21b
Fix warnings from new flags in cmakelists.
