+++ source/gwb-grid/main.cc
@@ -87,7 +87,7 @@ void filter_vtu_mesh(int dim,
                      std::vector<vtu11::DataSetData> &output_data)
 {
   output_data.resize(input_data.size());
-  const std::int64_t invalid = static_cast<std::uint64_t>(-1);
+  const std::int64_t invalid = -1;
   std::vector<std::int64_t> vertex_index_map(input_mesh.points().size(), invalid);
 
   const unsigned int n_vert_per_cell = (dim==3)?8:4;
@@ -99,25 +99,25 @@ void filter_vtu_mesh(int dim,
       int highest_tag = -1;
       for (size_t idx=cellidx*n_vert_per_cell; idx<(cellidx+1)*n_vert_per_cell; ++idx)
         {
-          const std::int64_t src_vid = input_mesh.connectivity()[idx];
+          const std::size_t src_vid = static_cast<size_t>(input_mesh.connectivity()[idx]);
           highest_tag = std::max(highest_tag,static_cast<int>(input_data[2][src_vid]));
         }
-      if (highest_tag < 0 || include_tag[highest_tag]==false)
+      if (highest_tag < 0 || include_tag[static_cast<size_t>(highest_tag)]==false)
         continue;
 
       ++dst_cellid;
 
       for (size_t idx=cellidx*n_vert_per_cell; idx<(cellidx+1)*n_vert_per_cell; ++idx)
         {
-          const std::int64_t src_vid = input_mesh.connectivity()[idx];
+          const size_t src_vid = static_cast<size_t>(input_mesh.connectivity()[idx]);
 
           std::int64_t dst_vid = vertex_index_map[src_vid];
           if (dst_vid == invalid)
             {
-              dst_vid = output_mesh.points().size()/3;
+              dst_vid = static_cast<std::int64_t>(output_mesh.points().size()/3);
               vertex_index_map[src_vid] = dst_vid;
 
-              for (int i=0; i<3; ++i)
+              for (unsigned int i=0; i<3; ++i)
                 output_mesh.points().push_back(input_mesh.points()[src_vid*3+i]);
 
               for (unsigned int d=0; d<input_data.size(); ++d)
@@ -127,7 +127,7 @@ void filter_vtu_mesh(int dim,
           output_mesh.connectivity().push_back(dst_vid);
 
         }
-      output_mesh.offsets().push_back(dst_cellid*n_vert_per_cell);
+      output_mesh.offsets().push_back(static_cast<long int>(dst_cellid*n_vert_per_cell));
 
       output_mesh.types().push_back(input_mesh.types()[cellidx]);
     }
