### 2024-06-06 cf60b94
more python tester output
