### 2024-06-06 e35d851
fix run_python_tests.cmake to get it to run on ubuntu 24.04.
