+++ tests/python/run_python_tests.cmake
@@ -20,7 +20,7 @@ set(EXECUTE_COMMAND ${TEST_PROGRAM} ${TEST_ARGS})
 # run the test program, capture the stdout/stderr and the result var ${TEST_ARGS}
 execute_process(
   COMMAND ${TEST_PROGRAM} ${TEST_ARGS} -i ${TEST_INPUT}  
-  WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/fortran/ 
+  WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/python/ 
   OUTPUT_FILE ${TEST_OUTPUT}
   ERROR_VARIABLE TEST_ERROR_VAR
   RESULT_VARIABLE TEST_RESULT_VAR
