### 2024-06-06 75a6018
Set requirements for running the clang tests.
