### 2024-06-08 b703c03
Adress comment Timo and change int to size_t in changes from #713
