### 2024-06-25 #6671 @busstoptaktik
Everything seems fine, although I have still not been able to compile the material on (Windows 10, gcc 14.1, cmake 3.24.3, ninja 1.11.1). I have, however, no reason to believe this is due to anything but a local glitch, so with this caveat, I have checked off the tick boxes in the "Functionality" section as well
