### 2024-06-26 #6671 @cpgr
Just confirming that I was able to build this project (OSX), and could reproduce the example in the paper, as well as run all of the tutorials
