### 2024-07-02 #6671 @observingClouds
@busstoptaktik, coming back to [your initial issues with the compilation](https://github.com/openjournals/joss-reviews/issues/6671?notification_referrer_id=NT_kwDOApl-tbQxMDM4MzkyOTc2ODo0MzYxMzg3Nw#issuecomment-2101245197), has this been resolved by you switching to a different machine or has there been some fixes being made? I would just like to get a feeling whether the compilation issue is across a broad range of platforms or if it was rather due to a potential "glitch" on one of your machines.

Thanks!
