### 2024-07-28 GeodynamicWorldBuilder/WorldBuilder#743 @MFraters
Windows MinGW support
The CI tester for windows MinGW was dropped in https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/287 (see pull request for details). Note that compiling with visual studio on windows still works and is tested.

While it back then still compiled manually, the doctest and vtu11 libraries use very low level system calls (such as _fileno) and do get confused by the MinGW/CYGWIN environment. Until this is fixed, we should no longer recomend the use of MinGW to compile the world builder on windows, but instead only recomend linux subsystems for windows or visual studio.

Background:
Based on a comment by @busstoptaktik during the review of the JOSS paper (https://github.com/openjournals/joss-reviews/issues/6671), where there where issues with compiling it with Windows 10, gcc 14.1, cmake 3.24.3, ninja 1.11.1, I inferred that this had to do with MinGW, so I tried it myself and ran into the issue described above.
