### 2024-07-28 18170f9
Deprecate windows mingw gcc compiler.
