### 2024-08-01 #6671 @busstoptaktik
> I would just like to get a feeling whether the compilation issue is across a broad range of platforms or if it was rather due to a potential "glitch" on one of your machines.

As noted by @MFraters [here](https://github.com/GeodynamicWorldBuilder/WorldBuilder/issues/743), this is a general limitation regarding compilation on the Windows platform, where only MSVC is supported, which is unfortunate, but not a show-stopper
