### 2024-08-01 9170568
fully remove MinGW support in documentation.
