### 2024-08-01 faab668
Update doc/sphinx/developer_manual/contributing_to_the_documentation/documentation_style_guide.md
