### 2024-08-16 #6671 @observingClouds
@busstoptaktik thank you for the additional comments and evaluation. As this only effects the windows platform I'm happy to move on with the submission, close the review process and work with @MFraters through the post-review steps before accepting the submission for publication.

@busstoptaktik, @cpgr thank you very much for your time and effort. We appreciate it!
