### 2024-08-16 #6671 @observingClouds
@MFraters could you check above's missing DOI? You mentioned in https://github.com/openjournals/joss-reviews/issues/6470#issuecomment-1998062987 that you would now be able to update the AGU poster reference.
