### 2024-08-16 #6671 @observingClouds
@MFraters reading through the paper I have a few more comments:
- could you please ensure that the capitalization of the references in the bibliography is preserved (Double `{{ }}` in paper.bib usually work)
- could you please change the referencing style of some references in parenthesis following https://mystmd.org/guide/citations#table-pandoc-citations; so that e.g. ```(Saxena et al. (2023), Gea et al. (2023), Sandiford & Craig (2023),46
and van der Wiel et al. (2024))``` is more similar to ```(Saxena et al., 2023; Gea et al., 2023; Sandiford & Craig, 2023
and van der Wiel et al., 2024)```
Thank you!
