+++ source/world_builder/features/oceanic_plate_models/composition/tian2019_water_content.cc
@@ -1,5 +1,5 @@
 /*
-  Copyright (C) 2018 - 2021 by the authors of the World Builder code.
+  Copyright (C) 2018 - 2024 by the authors of the World Builder code.
 
   This file is part of the World Builder.
 
@@ -17,7 +17,7 @@
    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-#include "world_builder/features/oceanic_plate_models/composition/water_content.h"
+#include "world_builder/features/oceanic_plate_models/composition/tian2019_water_content.h"
 
 #include "world_builder/kd_tree.h"
 #include "world_builder/nan.h"
@@ -28,6 +28,7 @@
 #include "world_builder/types/unsigned_int.h"
 #include "world_builder/types/value_at_points.h"
 #include "world_builder/utilities.h"
+#include "world_builder/world.h"
 
 
 namespace WorldBuilder
@@ -41,7 +42,7 @@ namespace WorldBuilder
     {
       namespace Composition
       {
-        WaterContent::WaterContent(WorldBuilder::World *world_)
+        TianWaterContent::TianWaterContent(WorldBuilder::World *world_)
           :
           min_depth(NaN::DSNAN),
           max_depth(NaN::DSNAN),
@@ -51,16 +52,23 @@ namespace WorldBuilder
           this->name = "water content";
         }
 
-        WaterContent::~WaterContent()
+        TianWaterContent::~TianWaterContent()
           = default;
 
         void
-        WaterContent::declare_entries(Parameters &prm, const std::string & /*unused*/)
+        TianWaterContent::declare_entries(Parameters &prm, const std::string & /*unused*/)
         {
           // Document plugin and require entries if needed.
           // Add compositions to the required parameters.
           prm.declare_entry("", Types::Object({"compositions"}),
-                            "WaterContent compositional model. Sets water content as a compositional field.");
+                            "TianWaterContent compositional model. Sets bound water content as a compositional field. The returned "
+                            "water content is based on the the temperature and pressure at a point within the world. Currently, "
+                            "the bound water content can be determined for four different lithologies: 'sediment', mid-ocean "
+                            "ridge basalt ('MORB'), 'gabbro', and 'peridotite', using parameterized phase diagrams from Tian et al., 2019 "
+                            "(https://doi.org/10.1029/2019GC008488). The pressure is lithostatic, calculated with a constant user defined "
+                            "density, and is limited by a user defined cutoff pressure (in GPa) for each lithology. This is required because the "
+                            "parameterization breaks down at large pressures. Recommended cutoff pressures are 10 GPa is used for 'peridotite', "
+                            "26 GPa is used for 'gabbro', 16 GPa is used for 'MORB', and 1 GPa is used for 'sediment'.");
 
           // Declare entries of this plugin
           prm.declare_entry("min depth", Types::OneOf(Types::Double(0),Types::Array(Types::ValueAtPoints(0., 2.))),
@@ -70,22 +78,26 @@ namespace WorldBuilder
           prm.declare_entry("compositions", Types::Array(Types::UnsignedInt(),0),
                             "A list with the labels of the composition which are present there.");
           prm.declare_entry("density", Types::Double(3000.0),
-                            "The reference density used to estimate the lithostatic pressure for calculating "
+                            "The reference density used for determining the lithostatic pressure for calculating "
                             "the bound water content.");
           prm.declare_entry("lithology",  Types::String("peridotite"),
-                            "The lithology used to determine which polynomials to use for calculating the water content.");
[…]
