### 2024-08-19 #6671 @MFraters
Thanks @busstoptaktik and @cpgr for your review and finding the issue with mingw builds, and of course thanks @observingClouds for your work and your further comments! I will go through the issues in the next few days. It will take me a bit to make the new release (as mentioned in your post-review checklist) since it requires carefully going through some steps.

> @MFraters could you check above's missing DOI? You mentioned in [#6470 (comment)](https://github.com/openjournals/joss-reviews/issues/6470#issuecomment-1998062987) that you would now be able to update the AGU poster reference.

Unfortunately the paper is not yet submitted. How do you want pro procede with this? The abstract can be found here: https://agu.confex.com/agu/fm23/meetingapp.cgi/Paper/1363473, but doesn't have a doi.

> @MFraters reading through the paper I have a few more comments:
>
>     * could you please ensure that the capitalization of the references in the bibliography is preserved (Double `{{ }}` in paper.bib usually work)
>
>     * could you please change the referencing style of some references in parenthesis following https://mystmd.org/guide/citations#table-pandoc-citations; so that e.g. `(Saxena et al. (2023), Gea et al. (2023), Sandiford & Craig (2023),46 and van der Wiel et al. (2024))` is more similar to `(Saxena et al., 2023; Gea et al., 2023; Sandiford & Craig, 2023 and van der Wiel et al., 2024)`
>       Thank you!

Yes, I will address there, thanks!
