### 2024-08-19 5f868cf
Update doc/sphinx/user_manual/basic_starter_tutorial/18_2D_models.md
