### 2024-08-19 cf7378e
Update doc/sphinx/user_manual/basic_starter_tutorial/08_passive_margin_variable_depth.md
