### 2024-08-19 GeodynamicWorldBuilder/WorldBuilder#746 @MFraters
Release v1.0
The JOSS paper is now in the post-review steps! See https://github.com/openjournals/joss-reviews/issues/6671 for more info.

This means that a new release is needed.

Here is the todo list:

Here is the todo list:
- [x] confirm with @observingClouds what to do with the doi issue on the agu abstract -> add url to reference (#747)
- [x] ensure that the capitalization of the references in the bibliography is preserved (Double {{ }} in paper.bib usually work) (see https://github.com/openjournals/joss-reviews/issues/6671#issuecomment-2293736704) (#748)
- [x]  change the referencing style of some references in parenthesis following https://mystmd.org/guide/citations#table-pandoc-citations; so that e.g. (Saxena et al. (2023), Gea et al. (2023), Sandiford & Craig (2023),46 and van der Wiel et al. (2024)) is more similar to (Saxena et al., 2023; Gea et al., 2023; Sandiford & Craig, 2023 and van der Wiel et al., 2024) (see https://github.com/openjournals/joss-reviews/issues/6671#issuecomment-2293736704) (#748)
- [x] Merge #729 and #719
- [x] Decide whether to merge #661 before or after release ~~(decide after review)~~ Merged!
- [x] run spell checker -> integrated in CI
- [x] run clang-analyzer scan-build (https://clang-analyzer.llvm.org/command-line.html) -- No bugs found.
- [x] run clang tidy
- [x] check and fix changelog entries (#749,#756)
- [x] Update copyright year sphinx, include and header files -> Year is still the same as the 0.6 release, so no need to upate
- [x] update schema and delclarations file (#754)
- [x] Confirm aspect pull request review (https://github.com/geodynamics/aspect/pull/6012)
- [x] make new zenodo doi (before the release) to include in the new release -> 10.5281/zenodo.13352667 (added in #750)
- [x] Get a doi for the JOSS paper ->  https://doi.org/10.21105/joss.06671 (#750)
- [x] Update readme with new release information (#757)
- [x] increase version number (https://github.com/GeodynamicWorldBuilder/WorldBuilder/pull/753)
- [x] make release
- [x] post-release version bump --> (#758) Note: Benchmarks won't work in this pull request, but will work after pull request is merged.

Please let me know if there are any other items which should be checked before the release, or which may delay it.
