### 2024-08-20 #6671 @observingClouds
> Unfortunately the paper is not yet submitted. How do you want pro procede with this? The abstract can be found here: https://agu.confex.com/agu/fm23/meetingapp.cgi/Paper/1363473, but doesn't have a doi.

It's okay if we just use the URL. We prefer of course DOIs but if there is none, we can use the url.
