### 2024-08-20 GeodynamicWorldBuilder/WorldBuilder#743 @MFraters
resolved by #287. We decided to drop support of  MinGW/CYGWIN given the alternatives for windows are generally better and more stable.
