+++ doc/JOSS/1.0/paper.bib
@@ -146,6 +146,7 @@ DOI = {10.5194/gmd-14-3899-2021}
   author={Billen, Magali I and Fraters, Menno},
   journal={AGU23},
   year={2023},
+  url={https://agu.confex.com/agu/fm23/meetingapp.cgi/Paper/1363473},
   publisher={AGU}
 }
 
