### 2024-08-20 3b3762d
Fix dash in JOSS paper bib.
