+++ doc/JOSS/1.0/paper.bib
@@ -59,7 +59,7 @@ authors = {Moresi, L. and Dufour, F. and Mühlhaus, HB. },
 title = {Mantle Convection Modeling with Viscoelastic/Brittle Lithosphere: Numerical Methodology and Plate Tectonic Modeling.},
 journal = {Pure appl. geophys.},
 volume = {159},
-pages = {2335–2356},
+pages = {2335-2356},
 year = {2002},
 doi = {10.1007/s00024-002-8738-3},
 url = {https://doi.org/10.1007/s00024-002-8738-3}
