### 2024-08-20 31ce7d9
Ensure correct capatalisation of references with double {}.
