+++ doc/JOSS/1.0/paper.bib
@@ -1,6 +1,6 @@
 @Article{Fraters_Thieulot_etal_2019,
 AUTHOR = {Fraters, M. and Thieulot, C. and van den Berg, A. and Spakman, W.},
-TITLE = {The Geodynamic World Builder: a solution for complex initial conditions in numerical modeling},
+TITLE = {{The Geodynamic World Builder: a solution for complex initial conditions in numerical modeling}},
 JOURNAL = {Solid Earth},
 VOLUME = {10},
 YEAR = {2019},
@@ -12,7 +12,7 @@ DOI = {10.5194/se-10-1785-2019}
 
 @article{Bauville_Baumann_2019,
 author = {Bauville, A. and Baumann, T. S.},
-title = {geomIO: An Open-Source MATLAB Toolbox to Create the Initial Configuration of 2-D/3-D Thermo-Mechanical Simulations From 2-D Vector Drawings},
+title = {{geomIO: An Open-Source MATLAB Toolbox to Create the Initial Configuration of 2-D/3-D Thermo-Mechanical Simulations From 2-D Vector Drawings}},
 journal = {Geochemistry, Geophysics, Geosystems},
 volume = {20},
 number = {3},
@@ -27,7 +27,7 @@ year = {2019}
 
 @article{Spang_Baumann_2022,
 author = {Spang, A. and Baumann, T. S. and Kaus, B. J. P.},
-title = {Geodynamic Modeling With Uncertain Initial Geometries},
+title = {{Geodynamic Modeling With Uncertain Initial Geometries}},
 journal = {Geochemistry, Geophysics, Geosystems},
 volume = {23},
 number = {6},
@@ -50,13 +50,13 @@ volume = {4},
 number = {36}, 
 pages = {1136}, 
 author = {Romain Beucher and Louis Moresi and Julian Giordani and John Mansour and Dan Sandiford and Rebecca Farrington and Luke Mondy and Claire Mallard and Patrice Rey and Guillaume Duclaux and Owen Kaluza and Arijit Laik and Sara Morón}, 
-title = {UWGeodynamics: A teaching and research tool for numerical geodynamic modelling}, 
+title = {{UWGeodynamics: A teaching and research tool for numerical geodynamic modelling}}, 
 journal = {Journal of Open Source Software} 
 } 
 
 @article{Moresi_Dufour_2002,
 authors = {Moresi, L. and Dufour, F. and Mühlhaus, HB. },
-title = {Mantle Convection Modeling with Viscoelastic/Brittle Lithosphere: Numerical Methodology and Plate Tectonic Modeling.},
+title = {{Mantle Convection Modeling with Viscoelastic/Brittle Lithosphere: Numerical Methodology and Plate Tectonic Modeling.}},
 journal = {Pure appl. geophys.},
 volume = {159},
 pages = {2335-2356},
@@ -67,7 +67,7 @@ url = {https://doi.org/10.1007/s00024-002-8738-3}
 
 @article{Varga_Schaaf_2019,
 AUTHOR = {de la Varga, M. and Schaaf, A. and Wellmann, F.},
-TITLE = {GemPy 1.0: open-source stochastic geological modeling and inversion},
+TITLE = {{GemPy 1.0: open-source stochastic geological modeling and inversion}},
 JOURNAL = {Geoscientific Model Development},
 VOLUME = {12},
 YEAR = {2019},
@@ -79,7 +79,7 @@ DOI = {10.5194/gmd-12-1-2019}
 
 @article{Schaaf_Varga_2021,
 AUTHOR = {Schaaf, A. and de la Varga, M. and Wellmann, F. and Bond, C. E.},
-TITLE = {Constraining stochastic 3-D structural geological models with topology information using approximate Bayesian computation in GemPy 2.1},
[…]
