+++ doc/JOSS/1.0/paper.bib
@@ -55,7 +55,7 @@ journal = {Journal of Open Source Software}
 } 
 
 @article{Moresi_Dufour_2002,
-authors = {Moresi, L. and Dufour, F. and Mühlhaus, HB. },
+author = {Moresi, L. and Dufour, F. and Mühlhaus, HB.},
 title = {{Mantle Convection Modeling with Viscoelastic/Brittle Lithosphere: Numerical Methodology and Plate Tectonic Modeling.}},
 journal = {Pure appl. geophys.},
 volume = {159},
