### 2024-08-20 #6671 @MFraters
@observingClouds I have addressed the comments about the paper. I am now working on getting the release out, but that may take a bit because I usually also confirm that the release is ready to merge in the projects that include it (In this case just ASPECT).

For the release it would be nice if I could already point to to this paper for people to cite. Is it possible to already get a doi reserved for the paper so that I can include it in the release?
