### 2024-08-20 #6671 @observingClouds
The DOI will be https://doi.org/10.21105/joss.06671. It won't be registered after publication though.
