### 2024-08-20 GeodynamicWorldBuilder/WorldBuilder#746 @tjhei
> * --> Note: Benchmarks won't work until this pull request is merged.

I would expect that you can bump GWB version and the expected version and the .json files in the same commit, no?
