### 2024-08-20 GeodynamicWorldBuilder/WorldBuilder#746 @MFraters
If I remember correctly, the problem is that the benchmarks test the main version of the world builder, against the pull request version of the world builder, with the pull request version input files (you want to use the same input files for both). So the pull request which bumps the version, the benchmarks won't work.
