### 2024-08-20 bedc154
Adress comment from Juliane Dannberg on JOSS paper.
