### 2024-08-21 1ed4b67
Bump world builder version to v1.0.0.
