### 2024-08-21 33da27a
Set changelog unreleased to version to 1.0.0
