### 2024-08-21 2a2c62d
Update world builder schema and delcarations.
