### 2024-08-27 f20ca1b
Fix changelog issue/pull request links.
