### 2024-08-28 25b2bed
Update readme for v1.0.0
