### 2024-08-28 cd245a3
Apply suggestions from code review Rene
Co-authored-by: Rene Gassmoeller <rene.gassmoeller@mailbox.org>
