### 2024-08-29 #6671 @MFraters
@observingClouds All your comments have been addressed and a release for version 1.0.0 has been made: https://github.com/GeodynamicWorldBuilder/WorldBuilder/releases/tag/v1.0.0
