### 2024-08-29 #6671 @observingClouds
@MFraters thanks for the update and release of version 1.0.0. Looking at the archive at [10.5281/zenodo.13352667](https://doi.org/10.5281/zenodo.13352667) I have three comment/questions:
- JOSS prefers if the archive title matches the manuscript title. Could you edit the metadata of the release and adapt the title?
- The list of authors of the archive does not match the authors of the manuscript. Could you add everyone to the release or explain why the co-authors are not included?
- the version of the archive seems off in some places. Instead of `v1.0.0` it is `v6`. Please fix this inconsistency.

Thank you!
