### 2024-08-29 #6671 @MFraters
Thanks for the quick reply!

>     * JOSS prefers if the archive title matches the manuscript title. Could you edit the metadata of the release and adapt the title?

Done

>     * The list of authors of the archive does not match the authors of the manuscript. Could you add everyone to the release or explain why the co-authors are not included?

Done

>     * the version of the archive seems off in some places. Instead of `v1.0.0` it is `v6`. Please fix this inconsistency.

Thanks for catching this. I forgot to set the version field. Fixed it.
