### 2024-08-29 #6671 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
