### 2024-08-29 #6671 @MFraters
Great, thanks for all your help and work on this paper!

A quick question, I would like to merge new things into the main branch, but I am wondering whether the bot is still attached to that branch or not. I have crated a separate branch (https://github.com/GeodynamicWorldBuilder/WorldBuilder/tree/GWB-v1.0.0) for the 1.0 release, which can be used if needed and there is also the v1.0.0 tag. If you would rather have me wait a few days with changing anything on the main branch, let me know.
