### 2024-08-30 GeodynamicWorldBuilder/WorldBuilder#746 @MFraters
Release fully done! :tada:
