### 2024-09-04 #6671 @kthyng
@MFraters If you would like to join JOSS as a reviewer, please sign up at https://reviewers.joss.theoj.org/!
